refactor: split post tests
